restructure meta tests
